// playwright/support/db/schema.ts
import type { Generated } from 'kysely'

export interface OrdersTable {
    id: Generated<string>
    order_number: string
    color: string
    wheel_type: string
    customer_name: string
    customer_email: string
    customer_phone: string
    customer_cpf: string
    payment_method: string
    total_price: number
    status: string
    created_at: Generated<Date>
    updated_at: Generated<Date>
    optionals: string[] | null
}

export interface Database {
    orders: OrdersTable
}

export type NewOrderRow = {
    order_number: string
    color: string
    wheel_type: string
    customer_name: string
    customer_email: string
    customer_phone: string
    customer_cpf: string
    payment_method: string
    total_price: number
    status: string
    optionals?: string[] | null
}